pyproject.toml
[tool.poetry]
name = "kamilla"
version = "0.1.0"
description = ""
authors = ["snipe <72265661+notsniped@users.noreply.github.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
